refactor plot_njt
