+++ malariagen_data/anoph/distance_params.py
@@ -11,7 +11,11 @@ distance_metric: TypeAlias = Annotated[
     "The metric to compute distance between genotypes in two samples.",
 ]
 
+default_distance_metric: distance_metric = "cityblock"
+
 nj_algorithm: TypeAlias = Annotated[
     Literal["rapid", "canonical"],
     "Neighbour-joining algorithm to use.",
 ]
+
+default_nj_algorithm: nj_algorithm = "rapid"
