+++ malariagen_data/anopheles.py
@@ -1,4 +1,3 @@
-import math
 import warnings
 from abc import abstractmethod
 from bisect import bisect_left, bisect_right
@@ -18,7 +17,6 @@ import plotly.graph_objects as go  # type: ignore
 import xarray as xr
 from numpydoc_decorator import doc  # type: ignore
 
-from malariagen_data.anoph import tree_params
 from malariagen_data.anoph.snp_frq import (
     AnophelesSnpFrequencyAnalysis,
     _add_frequency_ci,
@@ -30,12 +28,10 @@ from .anoph import (
     aim_params,
     base_params,
     dash_params,
-    distance_params,
     gplt_params,
     hapnet_params,
     het_params,
     ihs_params,
-    pca_params,
     plotly_params,
     xpehh_params,
 )
@@ -3094,392 +3090,6 @@ class AnophelesDataResource(
 
         debug("launch the dash app")
         app.run(**run_params)
-
-    @doc(
-        summary="""
-            Plot an unrooted neighbour-joining tree, computed from pairwise distances
-            between samples using biallelic SNP genotypes.
-        """,
-        extended_summary="""
-            The tree is displayed as an unrooted tree using the equal angles layout.
-        """,
-        parameters=dict(
-            center_x="X coordinate where plotting is centered.",
-            center_y="Y coordinate where plotting is centered.",
-            arc_start="Angle where tree layout begins.",
-            arc_stop="Angle where tree layout ends.",
-            edge_legend="Show legend entries for the different edge (line) colors.",
-            leaf_legend="Show legend entries for the different leaf node (scatter) colors and symbols.",
-            legend_sizing="Controls sizing of items in legends, either 'trace' or 'constant'.",
-        ),
-    )
-    def plot_njt(
-        self,
-        region: base_params.regions,
-        n_snps: base_params.n_snps,
-        color: plotly_params.color = None,
-        symbol: plotly_params.symbol = None,
-        metric: distance_params.distance_metric = "cityblock",
-        distance_sort: Optional[tree_params.distance_sort] = None,
-        count_sort: Optional[tree_params.count_sort] = None,
-        center_x=0,
-        center_y=0,
-        arc_start=0,
-        arc_stop=2 * math.pi,
-        width: plotly_params.fig_width = 800,
-        height: plotly_params.fig_height = 600,
-        show: plotly_params.show = True,
-        renderer: plotly_params.renderer = None,
-        render_mode: plotly_params.render_mode = "svg",
-        title: plotly_params.title = True,
-        title_font_size: plotly_params.title_font_size = 14,
-        line_width: plotly_params.line_width = 0.5,
-        marker_size: plotly_params.marker_size = 5,
-        color_discrete_sequence: plotly_params.color_discrete_sequence = None,
-        color_discrete_map: plotly_params.color_discrete_map = None,
-        category_orders: plotly_params.category_order = None,
-        edge_legend: bool = False,
-        leaf_legend: bool = True,
-        legend_sizing: plotly_params.legend_sizing = "constant",
-        thin_offset: base_params.thin_offset = 0,
-        sample_sets: Optional[base_params.sample_sets] = None,
-        sample_query: Optional[base_params.sample_query] = None,
-        sample_query_options: Optional[base_params.sample_query_options] = None,
-        sample_indices: Optional[base_params.sample_indices] = None,
-        site_mask: Optional[base_params.site_mask] = base_params.DEFAULT,
-        site_class: Optional[base_params.site_class] = None,
-        min_minor_ac: Optional[
-            base_params.min_minor_ac
-        ] = pca_params.min_minor_ac_default,
-        max_missing_an: Optional[
-            base_params.max_missing_an
-        ] = pca_params.max_missing_an_default,
-        cohort_size: Optional[base_params.cohort_size] = None,
-        min_cohort_size: Optional[base_params.min_cohort_size] = None,
-        max_cohort_size: Optional[base_params.max_cohort_size] = None,
-        random_seed: base_params.random_seed = 42,
-        inline_array: base_params.inline_array = base_params.inline_array_default,
-        chunks: base_params.chunks = base_params.native_chunks,
-    ) -> plotly_params.figure:
-        from biotite.sequence.phylo import neighbor_joining  # type: ignore
-        from scipy.spatial.distance import squareform  # type: ignore
-
-        # Normalise params.
-        if count_sort is None and distance_sort is None:
-            count_sort = True
-            distance_sort = False
-
-        # Compute pairwise distances.
-        dist, samples, n_snps_used = self.biallelic_diplotype_pairwise_distances(
-            region=region,
-            n_snps=n_snps,
-            metric=metric,
-            thin_offset=thin_offset,
-            sample_sets=sample_sets,
-            sample_query=sample_query,
-            sample_query_options=sample_query_options,
-            sample_indices=sample_indices,
-            site_mask=site_mask,
-            site_class=site_class,
-            min_minor_ac=min_minor_ac,
-            max_missing_an=max_missing_an,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            random_seed=random_seed,
-            inline_array=inline_array,
-            chunks=chunks,
-        )
-
-        # Construct tree.
-        with self._spinner("Construct neighbour-joining tree"):
-            dist_sq = squareform(dist)
-            tree = neighbor_joining(dist_sq)
-
-        # Load sample metadata.
-        df_samples = self.sample_metadata(
-            sample_sets=sample_sets,
-            sample_query=sample_query,
-            sample_query_options=sample_query_options,
-            sample_indices=sample_indices,
-        )
-        # Ensure alignment with pairwise distances.
-        df_samples = df_samples.set_index("sample_id").loc[samples].reset_index()
-
-        # Normalise color and symbol parameters.
-        symbol_prepped = self._setup_sample_symbol(
-            data=df_samples,
-            symbol=symbol,
-        )
-        del symbol
-        (
-            color_prepped,
-            color_discrete_map_prepped,
-            category_orders_prepped,
-        ) = self._setup_sample_colors_plotly(
-            data=df_samples,
-            color=color,
-            color_discrete_map=color_discrete_map,
-            color_discrete_sequence=color_discrete_sequence,
-            category_orders=category_orders,
-        )
-        del color
-        del color_discrete_map
-        del color_discrete_sequence
-
-        # Extract data for leaf colors.
-        if color_prepped is not None:
-            leaf_colors = df_samples[color_prepped]
-        else:
-            leaf_colors = None
-
-        # Lay out the tree.
-        _, df_leaf_nodes, df_edges = unrooted_tree_layout_equal_angle(
-            tree=tree,
-            color=color_prepped,
-            leaf_colors=leaf_colors,
-            center_x=center_x,
-            center_y=center_y,
-            arc_start=arc_start,
-            arc_stop=arc_stop,
-            count_sort=count_sort,
-            distance_sort=distance_sort,
-        )
-
-        # Join leaf data with sample metadata.
-        df_leaf_data = df_leaf_nodes[["x", "y"]].join(df_samples, how="inner")
-        df_leaf_data.sort_index(inplace=True)
-
-        # Force default color for lines.
-        if color_discrete_map_prepped is not None:
-            color_discrete_map_prepped[""] = "gray"
-
-        # Construct plot title.
-        if title is True:
-            title_lines = []
-            if sample_sets is not None:
-                title_lines.append(f"Sample sets: {sample_sets}")
-            if sample_query is not None:
-                title_lines.append(f"Sample query: {sample_query}")
-            title_lines.append(f"Genomic region: {region} ({n_snps_used:,} SNPs)")
-            title = "<br>".join(title_lines)
-
-        # Start building the figure.
-        fig1 = px.line(
-            data_frame=df_edges,
-            x="x",
-            y="y",
-            color=color_prepped,
-            hover_name=color_prepped,
-            hover_data=None,
-            color_discrete_map=color_discrete_map_prepped,
-            category_orders=category_orders_prepped,
-            render_mode=render_mode,
-        )
-        fig1.update_traces(
-            showlegend=edge_legend,
-        )
-
-        # Configure hover data.
-        hover_data = self._setup_sample_hover_data_plotly(
-            color=color_prepped, symbol=symbol_prepped
-        )
-
-        # Add scatter plot to draw the leaves.
-        fig2 = px.scatter(
-            data_frame=df_leaf_data,
-            x="x",
-            y="y",
-            color=color_prepped,
-            symbol=symbol_prepped,
-            color_discrete_map=color_discrete_map_prepped,
-            category_orders=category_orders_prepped,
-            render_mode=render_mode,
-            hover_name="sample_id",
-            hover_data=hover_data,
-        )
-        fig2.update_traces(
-            showlegend=leaf_legend,
-        )
-
-        fig = go.Figure()
-        fig.add_traces(list(fig1.select_traces()))
-        fig.add_traces(list(fig2.select_traces()))
-
-        # Style lines and markers.
-        line_props = dict(width=line_width)
-        marker_props = dict(size=marker_size)
-        fig.update_traces(line=line_props, marker=marker_props)
-
-        # Tidy up.
-        fig.update_layout(
-            title=title,
-            width=width,
-            height=height,
-            template="simple_white",
-            title_font=dict(
-                size=title_font_size,
-            ),
-            legend=dict(itemsizing=legend_sizing, tracegroupgap=0),
-        )
-
-        # Style axes.
-        fig.update_xaxes(
-            title=None,
-            mirror=False,
-            showgrid=False,
-            showline=False,
-            showticklabels=False,
-            ticks="",
-        )
-        fig.update_yaxes(
-            title=None,
-            mirror=False,
-            showgrid=False,
-            showline=False,
-            showticklabels=False,
-            ticks="",
-            # N.B., this is important, as it prevents distortion of the tree.
-            # See also https://plotly.com/python/axes/#fixed-ratio-axes
-            scaleanchor="x",
-            scaleratio=1,
-        )
-
-        if show:  # pragma: no cover
-            fig.show(renderer=renderer)
-            return None
-        else:
-            return fig
-
-
-def unrooted_tree_layout_equal_angle(
-    *,
-    tree,
-    color,
-    leaf_colors,
-    center_x,
-    center_y,
-    arc_start,
-    arc_stop,
-    distance_sort,
-    count_sort,
-):
-    # Set up outputs.
-    internal_nodes = []
-    leaf_nodes = []
-    edges = []
-
-    # Begin recursion.
-    _unrooted_tree_layout_equal_angle(
-        tree_node=tree.root,
-        leaf_colors=leaf_colors,
-        x=center_x,
-        y=center_y,
-        arc_start=arc_start,
-        arc_stop=arc_stop,
-        distance_sort=distance_sort,
-        count_sort=count_sort,
-        internal_nodes=internal_nodes,
-        leaf_nodes=leaf_nodes,
-        edges=edges,
-    )
-
-    # Load results into dataframes.
-    df_internal_nodes = pd.DataFrame.from_records(internal_nodes, columns=["x", "y"])
-    df_leaf_nodes = pd.DataFrame.from_records(
-        leaf_nodes, columns=["x", "y", "index", color]
-    ).set_index("index")
-    df_edges = pd.DataFrame.from_records(edges, columns=["x", "y", color])
-
-    return df_internal_nodes, df_leaf_nodes, df_edges
-
-
-def _unrooted_tree_layout_equal_angle(
-    tree_node,
-    leaf_colors,
-    x,
-    y,
-    arc_start,
-    arc_stop,
-    distance_sort,
-    count_sort,
-    internal_nodes,
-    leaf_nodes,
-    edges,
-):
-    if tree_node.children:
-        # Store internal node coordinates.
-        internal_nodes.append([x, y])
-
-        # Count leaves (descendants).
-        leaf_count = tree_node.get_leaf_count()
-
-        # Sort the subtrees.
-        if distance_sort:
-            children = sorted(tree_node.children, key=lambda c: c.distance)
-        elif count_sort:
-            children = sorted(tree_node.children, key=lambda c: c.get_leaf_count())
-        else:
-            children = tree_node.children
-
-        # Iterate over children, dividing up the current arc into
-        # segments of size proportional to the number of leaves in
-        # the subtree.
-        arc_size = arc_stop - arc_start
-        child_arc_start = arc_start
-        for child in children:
-            # Define a segment of the arc for this child.
-            child_arc_size = arc_size * child.get_leaf_count() / leaf_count
-            child_arc_stop = child_arc_start + child_arc_size
-
-            # Define the angle at which this child will be drawn.
-            child_angle = child_arc_start + child_arc_size / 2
-
-            # Access the distance at which to draw this child.
-            distance = child.distance
-
-            # Now use trigonometry to calculate coordinates for this child.
-            child_x = x + distance * math.sin(child_angle)
-            child_y = y + distance * math.cos(child_angle)
-
-            # Figure out edge colors.
-            edge_color = ""
-            if leaf_colors is not None:
-                edge_colors = set(
-                    leaf_colors.iloc[[leaf.index for leaf in child.get_leaves()]]
-                )
-                if len(edge_colors) == 1:
-                    edge_color = edge_colors.pop()
-
-            # Add edge.
-            # edges.append([x, child_x, y, child_y, edge_color])
-            edges.append([x, y, edge_color])
-            edges.append([child_x, child_y, edge_color])
-            edges.append([None, None, edge_color])
-
-            # Recurse to draw the child.
-            _unrooted_tree_layout_equal_angle(
-                tree_node=child,
-                leaf_colors=leaf_colors,
-                x=child_x,
-                y=child_y,
-                internal_nodes=internal_nodes,
-                leaf_nodes=leaf_nodes,
-                edges=edges,
-                arc_start=child_arc_start,
-                arc_stop=child_arc_stop,
-                count_sort=count_sort,
-                distance_sort=distance_sort,
-            )
-
-            # Update loop variables ready for the next iteration.
-            child_arc_start = child_arc_stop
-
-    else:
-        leaf_color = ""
-        if leaf_colors is not None:
-            leaf_color = leaf_colors.iloc[tree_node.index]
-        leaf_nodes.append([x, y, tree_node.index, leaf_color])
 
 
 @numba.njit("Tuple((int8, int64))(int8[:], int8)")
